003_meeting-note-bot-render/requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
notion-client>=2.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiohttp>=3.9.0
